Fix the package for C++14 compilers.
With C++14, the way exceptions are specified
changed, causing the build to fail. This was
most evident when using GCC 6+, as the default
standard used changed from C++11 to C++14.

Previously, GCC was using the C++98 standard
with -fexceptions, and required exception
specifiers. Using them with C++11, or newer
standards, now results in compilation errors.